requirements.txt
flask==0.10.1
flask-sqlalchemy==2.0
pymysql==0.7.1
sqlparse==0.1.18
odo==0.4.0
flask-restless==0.17.0
flask-admin==1.3.0
progressbar2==3.5.1
biopython==1.66
numpy==1.10.4
